refactor: extract schedule creation logic to custom hook
This moves schedule creation states and functions into a
`useCreateSchedule` hook for better organization.
